render: trim the pipeline-key branch after architecture review (#392)
- nt_gfx_pipeline_key lives in its own pure TU compiled into nt_gfx,
  nt_gfx_stub and the test fake: every gfx implementation carries the same
  public surface (module-layout rule); test_gfx_stub links and runs it.
- Drop the check_cache_keys.sh grep gate and its check.sh/CI hooks: it matched
  one constant and missed every other fold. The typed key and the neighbour
  tests are the guard; AGENTS.md says so in one bullet that points at the spec.
- Vertex-input builders no longer re-assert attr_map locations: the material
  boundary already did. The packer keeps its asserts (public, cache hits skip
  make_pipeline).
- Static-assert tripwire and spec describe what they catch: offset/size moves,
  not padding-hole insertions.
- dyn -> float_bits; key_equal compares bits then memcmp the floats.
- Comments cut to the 2-3 line WHY rule; color_mode range check uses an
  unsigned compare.
- Tests: -0.0 vs +0.0 splits pipelines; disabled-blend canonicalisation also
  dirties the ops.
